+++ files/Utilities/ReloadXML.app/Contents/MacOS/core.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-/Applications/KeyRemap4MacBook.app/Contents/Applications/KeyRemap4MacBook_cli.app/Contents/MacOS/KeyRemap4MacBook_cli \
+/Applications/Karabiner.app/Contents/Library/bin/karabiner \
     reloadxml
-/usr/bin/osascript -e 'display notification "Reloaded" with title "KeyRemap4MacBook"';
+/usr/bin/osascript -e 'display notification "Reloaded" with title "Karabiner"';
 
 exit 0